online_store_backend/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
Pillow==8.2.0
